Trick/requirements.txt
python-telegram-bot==20.6
cryptography~=46.0.3
google-generativeai
python-dotenv~=1.2.1